Feat: 사이드 바에 로그아웃 추가
